+++ fiddles/nativeScript/fiddle-0002-Puzz/watch.sh
@@ -1,3 +1,5 @@
-cd Puzz
-tns livesync android --emulator --watch
+nohup /Users/e13542/bin/android-sdk-macosx/tools/emulator -netdelay none -netspeed full -avd 320480QVGA &
+sleep 60;
+cd Puzz;
+tns livesync android --emulator --watch;